+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Fixed
+- Calculations of TVD resulted in an error when the distance between the two coordinates to interpolate was larger than their difference of MD.
+
 ## v2.1.1156 - 2025-04-04
 
 ### Added